+++ packages/compendium/common/index.ts
@@ -1,7 +1,8 @@
 export {
   searchResultTemplateEchostream,
   searchResultTemplateBazaario,
-  searchResultTemplateFoodAndWine
+  searchResultTemplateFoodAndWine,
+  searchResultTemplateGalaxyNews
 } from './searchcraft-search-results/search-result-template';
 export { customAdTemplate } from './searchcraft-ad/custom-ad-template';
 export {